Change NamespaceCreation opts.Align to be 64-bit value
This simplifies the use of Align variable, as most
of it's use, including use in downward API to libndctl,
uses 64 bit value, so we can get rid of many typecasts.
This also avoids integer overflow error that is caused
by using 32-bit value in intermediate variable.

+++ pkg/ndctl/ndctl.go
@@ -27,7 +27,7 @@ type CreateNamespaceOpts struct {
 	Name       string
 	Size       uint64
 	SectorSize uint64
-	Align      uint32
+	Align      uint64
 	Type       NamespaceType
 	Mode       NamespaceMode
 	Location   MapLocation
